requirements.txt
asgiref==3.8.0
coverage==7.4.4
Django==5.0.3
django-cors-headers==4.3.1
djangorestframework==3.15.0
djangorestframework-simplejwt==5.3.1
PyJWT==2.8.0
pytz==2024.1
sqlparse==0.4.4
